Add http request progress delegate 64 bit support.
[REVIEW] [at]michael.kirzinger [at]rafa.lecina [at]michael.atchison
#jira UE-182679
#rb [at]michael.atchison [at]rafa.lecina

[CL 26821886 by lorry li in ue5-main branch]

+++ Engine/Source/Runtime/Online/BackgroundHTTP/Public/GenericPlatform/GenericPlatformBackgroundHttpRequest.h
@@ -43,7 +43,7 @@ protected:
 	private:
 		
 		bool HasRetriesRemaining() const;
-		void UpdateHttpProgress(FHttpRequestPtr UnderlyingHttpRequest, int32 BytesSent, int32 BytesReceived);
+		void UpdateHttpProgress(FHttpRequestPtr UnderlyingHttpRequest, uint64 BytesSent, uint64 BytesReceived);
 
 		void CleanUpHttpRequest();
